+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `agents-sdk` will be documented in this file.
 
+## Unreleased
+
+### Added
+* Added `ConversationSerializer` class for serializing and deserializing conversation state
+
 ## v0.1.0 - 2025-03-20
 
 * Initial release of the agents-sdk